third_party/WebKit/LayoutTests/fast/dom/HTMLElement/insertAdjacentHTML-errors-expected.txt

Test insertAdjacentHTML exceptions to make sure they match HTML5

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS div.insertAdjacentHTML('beforeBegin', 'text') is undefined.
PASS div.insertAdjacentHTML('afterEnd', 'text') is undefined.
PASS div.insertAdjacentHTML('FOO', 'text') threw exception SyntaxError: An invalid or illegal string was specified..
PASS document.documentElement.insertAdjacentHTML('afterEnd', 'text') threw exception NoModificationAllowedError: An attempt was made to modify an object where modifications are not allowed..
PASS successfullyParsed is true

TEST COMPLETE
